refactor(api): split AuthHandler into 10 files by auth flow
Split auth_handler.go (1593 lines) into focused files:
- auth_handler.go: struct, constructor, core auth (signup/signin/signout/refresh)
- auth_handler_magiclink.go: magic link flow
- auth_handler_password.go: password reset flow
- auth_handler_email.go: email verification
- auth_handler_impersonation.go: impersonation flows
- auth_handler_mfa.go: TOTP/MFA setup and verification
- auth_handler_otp.go: OTP send/verify/resend
- auth_handler_identity.go: identity link/unlink, OIDC, reauth
- auth_handler_captcha.go: CSRF, captcha config/check
- auth_handler_config.go: auth config endpoint
